requirements.txt
fastapi
uvicorn[standard]
pymongo==3.11.1
motor
jose
pyjwt
passlib[bcrypt]
pydantic[email]